(PUP-1952) Fix hiera and lookup specs
Instead of trying to reload these files on every test from a different temporary
directory, which was not working due to caching in jruby, just load once from a
fixture.
